requirements.txt
2to3==1.0
alabaster==0.7.3
amqp==2.4.2
aniso8601==3.0.2
appnope==0.1.0
argh==0.26.2
astroid==1.3.6
attrs==19.1.0
Babel==1.3
backcall==0.1.0
backports-abc==0.5
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.shutil-which==3.5.1
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.6.3
billiard==3.6.0.0
bigcommerce==0.20.1
bleach==2.1.3
blessed==1.14.2
boto==2.46.1
boto3==1.7.40
botocore==1.10.84
cachetools==2.1.0
caniusepython3==7.0.0
celery==4.3.0
cement==2.8.2
certifi==2017.4.17
chardet==3.0.4
chargebee==2.7.3
click==6.7
cloudpickle==0.7.0
colorama==0.3.7
configparser==3.5.0
construct==2.5.3
coreapi==2.3.3
coreschema==0.0.4
coverage==4.0a5
cycler==0.10.0
dask==1.1.0
DateTime==4.3
decorator==4.3.2
defusedxml==0.6.0
distlib==0.2.8
Django==2.2.1
django-activity-stream==0.8.0
django-appconf==1.0.3
django-autocomplete-light==3.3.4
django-autoslug==1.9.4
django-bootstrap-form==3.4
django-bootstrap-forms==0.1
django-bower==5.2.0
django-columns==0.1.0
django-cors-headers==3.0.1
django-countries==5.3.3
django-extensions==2.1.7
django-filer==1.5.0
django-filter==2.1.0
django-formtools==2.1
django-googledrive-api==0.2.4
django-googledrive-storage==1.3.5
django-haystack==2.8.1
django-images==0.4.3
django-js-asset==1.2.2
django-jsonfield==1.2.0
django-localflavor==2.2
django-model-utils==3.2.0
django-mptt==0.10.0
django-multiupload==0.5.2
django-nose==1.4.6
django-notifications-hq==1.5.0
django-oauth-toolkit==1.2.0
django-oauth2-provider==0.2.6.1
django-pagination-py3==1.1.2
django-polymorphic==2.0.3
django-postman==3.6.2
django-render-block==0.6
django-rest-elasticsearch==0.4.1
django-sslify==0.2.7
django-sslserver==0.20
django-storages==1.7.1
django-tagging==0.4.6
django-templated-email==2.3.0
django-user-accounts==2.1.0
django-webpack-loader==0.6.0
djangorestframework==3.9.4
djangorestframework-datatables==0.5.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
drf-yasg==1.16.0
drfdocs==0.0.11
easy-thumbnails==2.3
elasticsearch==5.4.0
elasticsearch-dsl==5.3.0
entrypoints==0.2.3
enum34==1.1.6
et-xmlfile==1.0.1
future==0.17.1
futures==3.1.1
gcloud==0.17.0
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
googlemaps==2.2
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
haystack==0.42
html2text==2018.1.9
html5lib==1.0.1
httplib2==0.9.2
httplib2shim==0.0.3
httpretty==0.8.8
idna==2.7
inflection==0.3.1
ipdb==0.10.3
ipykernel==4.8.2
ipython==5.5.0
ipython-genutils==0.2.0
ipywidgets==7.2.0
itsdangerous==0.24
itypes==1.1.0
jdcal==1.3
jedi==0.13.3
Jinja2==2.7.3
jmespath==0.9.3
jsonfield==2.0.2
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jws==0.1.3
kiwisolver==1.0.1
kombu==4.5.0
lazy-object-proxy==1.3.1
livereload==2.5.1
lml==0.0.1
logilab-common==0.63.2
MarkupSafe==0.23
matplotlib==2.2.3
mistune==0.8.3
mock==1.0.1
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.2
nose==1.3.7
notebook==5.4.1
notifications==0.3.2
numpy==1.16.3
oauth2==1.9.0.post1
oauth2-provider==0.0
oauth2client==4.1.3
oauthlib==3.0.2
olefile==0.44
openpyxl==2.4.5
ordereddict==1.1
packaging==19.0
pam==0.1.4
pandocfilters==1.4.2
parso==0.4.0
pathlib2==2.3.0
pathspec==0.5.0
pathtools==0.1.2
pdfkit==0.6.1
pefile==2018.8.8
pexpect==4.4.0
phonenumbers==8.10.9
pickleshare==0.7.4
Pillow==5.4.1
pinax-theme-bootstrap==7.2.0
port-for==0.3.1
prices==1.0.0
prometheus-client==0.6.0
promise==2.2.1
prompt-toolkit==1.0.15
protobuf==3.7.1
psycopg2==2.7.3.2
psycopg2-binary==2.7.7
ptyprocess==0.5.2
py2-py3-django-email-as-username==1.7.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycryptodome==3.4.3
pydot==1.4.1
PyDrive==1.3.1
pyexcel==0.4.5
pyexcel-io==0.3.3
pyexcel-webio==0.0.11
pyexcel-xls==0.3.1
pyexcel-xlsx==0.3.0
Pygments==2.0.2
pyimage==0.1.1
PyJWT==1.5.3
pylabels==1.2.1
pylint==1.4.3
pyparsing==2.3.1
Pyrebase==3.0.27
pyrsistent==0.15.0
PySocks==1.6.8
pystache==0.5.4
pytest-runner==5.1
python-dateutil==2.6.1
python-jwt==2.0.1
python-oauth2==1.1.0
python-openid==2.2.5
python-ptrace==0.9.3
python-stdnum==1.11
python3-openid==3.1.0
PyTumblr==0.0.6
pytz==2019.1
PyWavelets==1.0.1
pywinpty==0.5.4
PyYAML==5.1.1
pyzmq==17.0.0
qrcode==6.1
qtconsole==4.3.1
reportlab==3.5.13
requests==2.22.0
requests-oauthlib==0.4.2
requests-toolbelt==0.7.0
rest-social-auth==2.1.0
rsa==3.4.2
ruamel.yaml==0.15.99
Rx==1.6.1
s3transfer==0.1.10
scikit-image==0.14.2
scipy==1.2.0
semantic-version==2.5.0
semantics3==0.3.6
Send2Trash==1.5.0
shortuuid==0.5.0
simplegeneric==0.8.1
simplejson==3.6.5
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
sorl-thumbnail==12.3
Sphinx==1.3.1
sphinx-autobuild==0.7.1
sphinx-rtd-theme==0.1.7
sqlparse==0.3.0
stripe==2.27.0
subprocess32==3.5.3
swapper==1.1.0
tabulate==0.7.5
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
texttable==0.8.8
tldr==0.4
toolz==0.9.0
tornado==4.5.1
traitlets==4.3.2
twilio==3.8.0
typed-ast==1.3.4
typing==3.6.6
Unidecode==0.4.21
uritemplate==3.0.0
urllib3==1.21.1
urltools==0.3.2
vine==1.3.0
virtualenv==15.1.0
watchdog==0.8.3
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.46.0
Werkzeug==0.14.1
widgetsnbextension==3.2.0
win-unicode-console==0.5
WooCommerce==1.2.1
wrapt==1.11.1
xlrd==1.0.0
xlwt==1.2.0
zenpy>=2.0
zope.interface==4.6.0

opencv-python~=4.2.0.34
pprint~=0.1
natsort~=7.0.1
removebg~=0.3